requirements.txt
# Core Dependencies
pywebview>=4.0.0
PyQt5>=5.15.0

# Screen Capture & Image Processing
mss>=9.0.0
Pillow>=10.0.0

# OCR
pytesseract>=0.3.10

# Translation & NLP
openai>=1.0.0
requests>=2.31.0

# Text Processing & NLP
scikit-learn>=1.3.0
nltk>=3.8.0

# File Processing
unstructured>=0.10.0
unstructured[pdf]
PyMuPDF>=1.23.0
python-docx>=0.8.11

# Configuration & Environment
python-dotenv>=1.0.0
cryptography>=41.0.0
keyring>=25.7.0

# Database
# sqlite3 is included in Python standard library

# Input Monitoring
pynput>=1.7.6
